backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.23,<1.5.0
pydantic>=1.8.2,<2.0.0
pdf2docx>=0.5.5,<0.6.0
python-docx>=0.8.11,<0.9.0
simplify-docx>=0.1.0,<0.2.0
spacy>=3.0.0,<4.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
aiofiles>=0.7.0,<0.8.0
python-dotenv>=0.19.0,<0.20.0 